Move all Faqs section component into one directory
