+++ next/src/components/layouts/Sections.tsx
@@ -12,7 +12,7 @@ import ContactsSection from '@/src/components/sections/ContactsSection'
 import DividerSection from '@/src/components/sections/DividerSection'
 import EventsSection from '@/src/components/sections/EventsSection'
 import FacilitiesSection from '@/src/components/sections/FacilitiesSection'
-import FaqsSection from '@/src/components/sections/FaqsSection'
+import FaqsSection from '@/src/components/sections/FaqsSection/FaqsSection'
 import FileListSection from '@/src/components/sections/FileListSection'
 import GallerySection from '@/src/components/sections/GallerySection'
 import IframeSection from '@/src/components/sections/IframeSection'
